#15 test case 수정
